Added sound effect
Added the Rare Item Get piano fanfare sound effect from The Legend of Zelda: Breath of the Wild. You can use python libraries to play back sounds when a cell finishes running. This is helpful as it allows you to be notified when a longer process finishes without having to stare at your screen nonstop.